+++ app/src/main/java/com/uzlov/valitova/justcargo/repo/datasources/IDeliveryRemoteDataSource.kt
@@ -7,7 +7,7 @@ interface IDeliveryRemoteDataSource {
     fun getDelivery() : LiveData<List<Delivery>>
     fun getDelivery(id: Int) : LiveData<Delivery?>
     fun putDelivery(delivery: Delivery)
-    fun removeDelivery(id: Int)
+    fun removeDelivery(id: Long)
     fun getDelivery(id: Long, phone: String): LiveData<Delivery?>
     fun getDeliveriesWithCarrierPhone(phone: String): LiveData<List<Delivery>>
 }
