+++ app/src/main/java/com/uzlov/valitova/justcargo/viemodels/RequestsViewModel.kt
@@ -20,13 +20,13 @@ class RequestsViewModel @Inject constructor(private var requestsUseCases: Reques
     }
 
     // возвращает заявку с указанным ID
-    fun getRequest(id: Int) = requestsUseCases?.getRequests(id)
+    fun getRequest(id: Long) = requestsUseCases?.getRequests(id)
 
     // добавляет заявку
     fun addRequest(request: Request) = requestsUseCases?.putRequest(request)
 
     // удаляет заявку
-    fun removeRequest(id: Int) = requestsUseCases?.removeRequests(id)
+    fun removeRequest(id: Long) = requestsUseCases?.removeRequests(id)
 
     // возвращает все заявки с указанным статусом
     fun getRequestsWithStatus(status: Int) = requestsUseCases?.getRequestsWithStatus(status)
